does it work? Editor can't edit null
